requirements.txt
asgiref==3.3.1
Django==3.1.6
django-colorfield==0.4.1
django-nested-admin==3.3.3
Pillow==8.1.0
python-monkey-business==1.0.0
pytz==2021.1
qrcode==6.1
six==1.15.0
sqlparse==0.4.1
